+++ client/src/utils/string.ts
@@ -1,2 +1,0 @@
-export const firstLetterUppercase: (str: string) => string = (str) =>
-  str.charAt(0).toUpperCase() + str.slice(1);